处理资源泄露问题
