+++ crawlo/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '1.4.6'
+__version__ = '1.4.7'
